requirements.txt
# ─── API Framework ───────────────────────────────────────────
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9

# ─── LLM & Embeddings ────────────────────────────────────────
openai==1.30.1
tiktoken==0.7.0

# ─── Vector Stores ───────────────────────────────────────────
faiss-cpu==1.8.0
pinecone-client==3.2.2

# ─── Document Processing ─────────────────────────────────────
pypdf==4.2.0
python-docx==1.1.2
unstructured==0.14.4

# ─── Retrieval ───────────────────────────────────────────────
rank-bm25==0.2.2

# ─── Data Validation ─────────────────────────────────────────
pydantic==2.7.1
pydantic-settings==2.2.1

# ─── Observability ───────────────────────────────────────────
langsmith==0.1.65

# ─── Utilities ───────────────────────────────────────────────
numpy==1.26.4
python-dotenv==1.0.1
httpx==0.27.0
tenacity==8.3.0
rich==13.7.1

# ─── Testing ─────────────────────────────────────────────────
pytest==8.2.0
pytest-asyncio==0.23.7
pytest-cov==5.0.0
respx==0.21.1
